Interactive_Mode_Functional_In_Simulator
I misunderstood where the origin was so cubes where placed in incorrect positions. also the cubes lost collision for some reason so i had to put collision back on them. it is still very hard to move the cubes in the simulator as its not meant for simulator, its meant for vr headset and controllers. it might be easier to move in simulator with one hand as sometimes two cubes will be grabed at once which should not happen. but everything should look normal now